Port supports multiple IP addresses
This changes the topology model for a port to allow multiple
IP addresses/subnets for a router port. This information is
encoded in the `port_subnet` field of the port message, which
changes from optional to repeated.

In addition, the patch changes the port simulation object
such that the addresses of a router port are represented by
a single list of IP subents. Each IP subnet will encode both
the address and subnet information. The class keeps two
portAddress4 and portAddress6 fields to allow simulation code
easy access to the first IPv4 and IPv6, respectively, minimizing
the number of changes. Each field is an IP subnet that encodes
both the port address and subnet. If a port does not have an IPv4
or IPv6, the corresponding field will be null. This requires the
simulation code to validate the port as an IPv4 or IPv6.

Backend compatibility is ensured in the following ways.

* Forward compatibility: MidoNet 5.3+ can read ports that were
  created in a previous version, since the Protocol Buffers
  message is compatible, and the port address and subnet fields
  will be merged when the port is loaded in the virtual topology.

* Backward compatibility: When creating a port, MidoNet 5.3+
  will encode both the address and subnet fields in the port
  message. These allows previous versions of MidoNet to load
  newer ports. This patch does not change the way Neutron
  translators or the REST API set the ports address and subnet,
  and therefore previous versions of MidoNet will receive ports
  with the same fields. Backward compatibility is not supported
  for ports with multiple addresses, which cannot be loaded in
  previous versions due to incompatibility of the message encoding
  (ZOOM will not be able to decode the repeated port_subnet value
  into an optional field).

API compatibility is ensured in the following ways.

* REST API requests that create/update router ports will specify
  the port address and subnet in the same way. This information
  is now converted to a repeated field. Requests for reading will
  work in the same way for ports having a single IP address.
  Ports with multiple IP addresses will return only the first
  address/subnet.

* REST API updates of ports with multiple IP addresses will
  overwrite the entire list of IP addresses and change the port
  to having a single IP address.

* Neutron translations will create the same port addresses and
  subnets as previous versions of MidoNet. The only changes
  affect the creation of container router ports and HA proxy
  ports, for which the subnet now includes the port address.

Note: Support to the REST API for ports with multiple addresses
will be added in a later patch.

Ref: MI-1778

Change-Id: Ia8aa301a6fa728a11768e55c68516b161ea601b1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>