Create Fip64 tunnel port from datapath controller
It turns out that to have upcalls for packets coming through a port, it
has to be created by the datapath controller (or rather, a netlink
connection that is actually monitoring for upcalls). So we have to
create the tunnel port on boot, and not on demand, which isn't as nice,
but it works.

Change-Id: I71c3006b89cf49da7740faecd84879cf608abffe
Signed-off-by: Ivan Kelly <ivan@midokura.com>