QOS: Interpret burst of 0 as default (80%)
The QosPolicyMapper already defaults to 80% of the sustained rate when the burst
is unset, but for some reason it always gets set to zero if the user doesn't
specify anything. Since a burst of zero doesn't make sense, this patch treats
zero as unset and defaults to 80% of the sustained rate.

Change-Id: Iaf6f1804cfb80d18f91ca36718e0ba1347ac6811
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-1790