Fix: Incorrect gateway field in local routes
When creating a router port, a route to the port address is added
automatically. This route is local but has the gateway field set
to 255.255.255.255, which is ignored during simulation but
displayed in the CLI.

This patch fixes the translator so that this unnecessary field
is not added anymore.

Fixes: MI-1793

Change-Id: I5a4dc59baf96c4df8d67c2d1f62cf0f78867703b
Signed-off-by: Adrian Serrano <adrian@midokura.com>
(cherry picked from commit 7393f1951cc852da28792a50c1fb252c29a8df54)