Router translator supports ports with multiple addresses
This allows the router translator to set the gateway to a network
with multiple subnets. For this purpose, the patch divides the
gateway port creation in three steps:

* Create a port builder with common fields, such as identifier,
  MAC, peer, administrative state.

* For each fixed IP address, configure the port address and
  subnet as follows: (i) the subnet is added to the list of port
  subnets and will include both the address and prefix length,
  (ii) the first address is added to the port for backward
  compatibility, (iii) if the address is IPv6, add a link-local
  IPv4 address to connect the port with the VPP downlink.

* After adding the new port to the transaction, for each port
  address create the port routes and rules as follows: (i) IPv4
  addresses routes and SNAT rules, (ii) IPv4 link-local route
  to the VPP downlink for IPv6 addresses, and NAT64 rules.

An important change is the mapping of port routes. Previously,
these had deterministic identifiers, since every port could have
at most one local route, port route, default gateway route.
These identifiers were used by the translator to delete the routes
when the ports were deleted.

However, when a port has multiple address, the port will have
several local, port, routes, etc. For this reason, the
deterministic algorithm to compute the route identifiers has
changed to include the address hash for which the route is added.

Furthermore, the port deletion does no longer delete the routes
explicitely by identifier, instead it relies on the ZOOM
referential integrity to remove them. This capability already
existed in previous versions of MidoNet and therefore the explicit
deletion was redundant.

Given that in previous versions of MidoNet deletion of port
routes is idempotent, the change should ensure compatibility,
that is:

* MidoNet 5.3+ should handle the routes deletion for ports created
in previous version.

* Previous versions should handle routes deletion for ports
created with MidoNet 5.3+ (actual routes are deleted by the ZOOM
referential integrity, and non-existing routes are ignored on
deletion).

Ref: MI-1776

Change-Id: I20bdb508c56a4842b050489df8c092a1fca8b728
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>