BGP routing uses the simulation port
Currently we use two different Route classes: a simulation
layer3.Route class and a legacy data.Route class that was kept
from the MidoNet 1.x data models.

For some reason, the BGP routing handles store routes using the
latter and then convert it back-and-forth to the simulation
version. This patch removes this overhead by removing the usage
for the legacy class, and using only the simulation verstion.

The patch also makes the following changes to Route.equals():

* Primitive members are compared with ==, rather than Object.equals
to prevent the boxing and unboxing.

* Routes should compare equal regardless of their weight or
attributes. The reason for this is that routes stored by Quagga
do not provide weight or attributes.

Change-Id: Ie6facd1a9c4db77c4d63d400b2c183e4be7bb8d0
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>