Fixes multi-address port compatibility
When loading in MidoNet 5.3+ a port that was created in a previous
version, the port addresses will display only the network part of
the address.

This fixes the problem by ensuring that if the complete subnet
does not match, the address list is updated accordingly.

The patch also adds unit tests.

Change-Id: I0116131a712e09067f0cd33482346bb6a74304fe
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>