Neutron API logs errors
This ensures that Neutron API errors are logged at debug level
to enable us to troubleshoot translator errors. The level is
debug because normally these errors are reflected in the REST
response.

Also, the patch removes the @Deprecated annotation and sets the
request timeout to the one specified in the configuration.

Change-Id: Ib9a9f9e0b38cc7ede0dd67bbc097b10cce2b00c8
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>