FIP64: Router mapper ignores IPv6 routes
This allows the router mapper to handle ports and routers with
IPv6 routes, by ignoring them. The mapper still monitors those
routes objects to ensure that the mapper does not enter a blind
spot where, waiting to load the routes, no updates are emitted.

To this end, the mapper adds a singleton representing an IPv6
route, which is ignored when generating notifications. This
requires minimum changes in the mapper state machine, and
guarantees that notifications are processed even for IPv6 routes.

Ref: MI-1815

Change-Id: I5041af43e17321401774b9fa25f6a2f1c090980e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>