Make QoS service logging less verbose
This downgrades certain log messages from info to debug and from
error to warning in the QoS service. This applies to notifications
received for topology changes, which we normally log only at
debug to reduce verbosity.

Also the patch changes the access modifiers of most methods to
private or protected.

Change-Id: Ifeeb8b02df3be6ecf5a08424984bd300317e5456
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>