QoS service uses scheduled executor
When creating an RX scheduler from a Java executor, if the
executor is not a ScheduledExecutorService, the RX will wrap the
call through an additional internal executor that allows tasks
to be submitted for arbitrary delays.

See: rx.schedulers.ExecutorScheduler.schedule()

To remove this overhead, this patch changes the QoS service
to use a ScheduledExecutorService by creating one with an
existing utility method. This also gives the executor thread a
distinct name, makes the thread a daemon thread, and sets it
shutdown policy to not throw exceptions if the executor is
shut down before all tasks are completed.

In addition, the QoS service shuts down the executor on stop,
and ensures that timer notifications are emitted on the scheduler
thread during retries.

Change-Id: I658ab2e2956817289f4bd166ce38cbee5298236c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>