FIP64: Router interface translator multiple subnets
This adds support for multiple subnets to the router interface
translator. For this purpose, the patch divides the creation of
a router interface into three steps:

* Create a port builder with common fields, such as identifier,
  MAC, peer, administrative state.

* For each fixed IP address, configure the port address and
  subnet as follows: (i) the subnet is added to the list of
  port subnets and will include both the address and prefix
  length, (ii) the first address is added to the port for
  backward compatibility, (iii) if the port is not uplink and
  is a gatway port, set its DHCP identifier.

Note: Since the IPv6 downlink ports no longer use simulation
through a local-link subnet, this subnet is no longer added to
the port.

* If the Neutron port does not have any fixed IP address,
  the port addresses are taken as the gateway IP addresses from
  the list of subnets specified in the router interface.

* If the port is uplink, add the port to the port group, set
  its tunnel key to a legacy tunnel key, and bind the port to
  the host-interface. For non-uplink ports, set the port peer,
  port group and if the port has at least one IPv6 address,
  set its tunnel key to a FIP64 tunnel key.

* After adding the new port to the transaction, for each port
  address create the port routes and rules as previously for
  both IPv4 and IPv6 addresses.

An important change is the maping of port routes. Previously,
these had deterministic identifiers, since every port could have
at most one local route, port route, default gateway route.
These identifiers were used by the translator to delete the routes
when the ports were deleted.

However, when a port has multiple address, the port will have
several local, port, routes, etc. For this reason, the
deterministic algorithm to compute the route identifiers has
changed to include the address hash for which the route is added.

Furthermore, the port deletion does no longer delete the routes
explicitely by identifier, instead it relies on the ZOOM
referential integrity to remove them. This capability already
existed in previous versions of MidoNet and therefore the explicit
deletion was redundant.

Given that in previous versions of MidoNet deletion of port
routes is idempotent, the change should ensure compatibility,
that is:

* MidoNet 5.3+ should handle the routes deletion for ports created
in previous version.

* Previous versions should handle routes deletion for ports
created with MidoNet 5.3+ (actual routes are deleted by the ZOOM
referential integrity, and non-existing routes are ignored on
deletion).

Change-Id: I1d28e462ab17d26574b11db6ea0c438eb55959f1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>