Validate packet cookie before restarting workflow
This adds the packet cookie to the RestartWorkflow class, such
that a packet context can be validated agains the cookie when
restarting the workflow on a postponed packet context.

This ensures that if the workflow has already completed for the
given context and the context was returned to the pool, we don't
restart the workflow for an invalid context.

This check was previously only made when adding the RestartWorkflow
message to the backchannel, but this did not guarantee that the
workflow won't be completed before the restart message will be
handled by the simulation.

Fixes: MI-1768

Change-Id: If659e00b7a65414c0bbafe40dcc0877d671e0fc1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>