Fix creation of child L4LB resources
This fixes the creation of child L4LB resources where the presence
of certain fields is not required, and currently not implemented
by the midonet-cli. This applies as follows:

* the pool identifier for a pool member

* the pool identifier for a VIP

To this end, the patch introduces a new validation filter, that
allows select API resources to prepare the DTOs before validation.
The rationale is that the API must still validate the fields when
the requests are made at the root URI.

Fixes: MI-1546

Change-Id: I2f9309e1d40af2326abdff77d6c54fd86c959aa6
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>