FIP64: Use config MAC only for FIP64 packets
This ensures that the post routing action of a router only
applies the FIP64 MAC to packets that match the NAT64 rule.
The IPv4 packets, even when egressing the router throgh the same
port should use the virtual topology MAC of the next hop, or
perform an ARP if the next hop is a phsyical one.

Change-Id: I9f78bb65ca45cefc283650c6bccb6ae57d1811ba
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>