Datapath actions to handle FIP64 state control
FIP64 functionality needs VPP service in the gateways to exchange
information between them about the usage of the IPs in source IPv4
pools.

This information exchange is initiated by a VPP service once an IP
is used from the pool, sending an update packet through the VPP
downlink encapsulated in a specific VNI (0x2ffffe).

OVS will ask Midolman about the rules for this flow. This patch add
the needed code to install a rules in OVS to forward these kind of
packets to all the gateways, chanding the VNI to (0x2fffff). Also,
a rule is installed to fordward it to the local VPP again (can be
used as an echo or ignored by the local VPP service).

In the receiving gateways, OVS will ask Midolman about the rules for
the flow with this 0x2fffff VNI, and Midolman will identify it as a
FIP64 tunnel (first 4 MSB bits of VNI are 0x2), forwarding it to the
local VPP service. This is fulfilled with the current implementation
so no further change is needed for this.

Also, definitions for these special tunnel keys has been added.

Ref: MI-1813

Change-Id: I4501ff0fdf4fe1fed92e05927c908327e9cb0117