Add ip_address to LB and session persist to Pool
These are only needed for lbaasv2 translations.  They will be
used to fill the VIP object with correct information, and
are harmless for any lbaas v1 or pre-existing LB objects.

Only the protobufs has changed, meaning the rest_api will not
have these fields viewable or changeable, as they are only needed
during the translation stage for LB, pool, and listeners.
Otherwise, they are completely internal.

Change-Id: Iaf811db18dabaca565f0d51f5366f347591d6a70