Session persistence in Pool
In order to match the LBaaS v2 models, this patch introduces session
persistence field to Pool.  The session persistent field in Vip still
remains for backward compatibility.  The new field is meant to be set
only by the C3PO translators, not REST API.  Session persistence is
applied if it is set on either Vip or Pool.

This patch also includes updating the pool member weight from 0 to 1
in the fake data generated in HealthMonitorConfigWatcherTest since
without it the tests do not pass.

Partially-Implements: MI-1798
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>

Change-Id: Ied3cd101f9617a797f44cd7932f463e41df80b6e