Hot fix for LoadBalancerV2IT test
After the change went in to change the IT tests to directly call
the translators, rather than asynchronously going through the
NeutronImporter, the LoadBalancerV2IT test started to fail, as it
had not been fixed due to the timing of the merges.

Instead of using sleeps and "eventually"s, this test can now
directly call the insert*Task methods and expect that the
translation will have been fully completed by the next function
call.  This makes the checks for error conditions a TON easier,
as we can just examine the exception being thrown by the
translation, rather than only inspect the presence or absence
of the objects.  But, since we weren't catching exceptions before,
as we were still assuming the asynch model, the new model caused
exceptions to be thrown directly by the test, which, because they
were not caught, would "fail" the test, despite being the correct
behavior.

So, changing the test to properly catch and examine these
exceptions makes the test pass, as well as making the test more
complete and precise.

Change-Id: Ic727973770b2e20db0e7c8d7b617ee8e54646379