Introduce classes for DHCP simulation objects
This are intended to replace the legacy DHCP data objects,
carried from MidoNet 1.x, and to allow the conversion from the
Protocol Buffers messages using ZOOM annotations. This will
allow in a subsequent patch to replace the multiple calls to
storage to load the DHCP objects with a single call.

Change-Id: I1609301dd1c099872665b9f6f67c6058e730d739
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>