FIP64: Allocate NAT pool between multiple uplink ports
This introduces an algoithm to allocate a NAT64 address pool
between multiple uplink ports. The algorithm works in a distributed
way, if all gateways have the same set of uplink ports.

The allocation method receives the complete NAT pool, the set of
uplink ports, and the identifier of the uplink port for which an
allocation should be made. It ensures that given the same set of
ports, even in different order, a given port will always receive
the same range, and the range is disjoint with the ranges that
will be allocated for the remanining ports.

Ref: MI-1781

Change-Id: Icb990c752b335b447f259cccb6fd65e696e5f34b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>