MDTS test to verify FIP64 LRU behavior
This patch introduces a new MDTS test for IPv6: test_lru. It creates
a server behind a FIP64 and as many TCP connections as the size of the
LRU are established from different IPv6 addresses. Then it verifies
that adding a new IPv6 client takes the least recently address.

It is disabled by default until LRU is enabled into VPP.

Change-Id: I91a30a26da0c50b8c174f685a063097aa417298e
Signed-off-by: Adrian Serrano <adrian@midokura.com>