Add vxlan tunnel setup to hashmap
Put statement was missing, causing rollback()
to fail on fib deassociation. If the fip6 is the last one
for this tenant VppController removes downlink. When new fib6 is created for
this tenant we will execute Vxlan setup again, but it will fail
since old vxlan tunnel still present in vpp.
Signed-off-by: Dmitry Bufistov <dmitry@midokura.com>

Change-Id: I25b8a1fe41e23b774df59047400f13078e285435