FIP64: Introduce a VPP uplink trait
This trait consolidates the management of the local uplink ports.
It monitors the local uplink port, and for each, it loads the port
data (including the IP address and port group membership). If the
port belongs to one or more stateful port groups, it loads the
corresponding port groups to determine the group membership.

Once the uplink data has been loaded from storage, it sends a
single notification that will permit the VPP controller to
configure the uplink.

If any of the relevant uplink port data changes, the trait will
emit a delete notification to remove any VPP configuration before
reconfiguring the VPP with the new setup.

Ref: MI-1783

Change-Id: Ie9786ac6aa23195f8f4dfb354287dc6433d8abcf
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>