FIP64: VPP controller uses uplinks trait
This replaces the uplink management logic from the VPP controller
with the one from the VPP uplink trait. This allows the controller
to handle a single notification stream for uplink changes,
including those that require a reconfiguration of the uplink when
relevant parameters, e.g. the IPv6 address or the stateful port
group membership, has changed.

Change-Id: I746b86060ba9ab1efade6bf1c8fc641332890935
Ref: MI-1783
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>