Replace DhcpConfig with single storage call
During the transition between MidoNet 1.x and 5.x, the DhcpConfig
trait was intended to allow the usage of dual-storage, by
abstracting the lower-level models. However, this abstraction is
inefficient for 5.x, since it performs multiple reads for the same
object.

Maintaining these classes is no longer needed. To this end, this
patch replaces the DHCP objects used by the DHCP implementations
with the versions introduced by an earlier patch, simplifies
their loading from storage, and removes the dead code.

Change-Id: Ia5e820e83de835a7c7854fe000815bc0327b89c0
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>