QoS: add DSCP mark to encapsulated packet
This patch simply sets the TOS field of the encapsulated (outer IP
header) packet to the DSCP mark specified by the QoS policy of the
associated port. The DSCP mark is the same as the one specified in
the inner IP header. The TOS in the tunnel packet is set when
translating a ToOutputPort virtual action so only tunneled packets
match.

This behavior is configurable through mn-conf and disabled by default.

Ref: IIOT-187
Change-Id: I2c7f0ca6095b694b1a90b2ec2ed888b7ea8a3ca0
Signed-off-by: Xavi León <xavi.leon@midokura.com>