Watchdog terminates only main agent process
Currently the MidoNet agent watchdog sends a SIGTERM to all
processes in the agent's progress group. This can be problematic,
since certain processes may exit while the agent is still
performing cleanup operations that may depend on them.

Furthermore, during a graceful exit, the agent is responsible for
stopping the child processes and thefore sending a SIGTERM to
all of them is not required.

To allow for a graceful shutdown, where the agent is able to
complete shutdown tasks that depend on spawned child processes,
this changes the behavior of the watchdog to send a SIGTERM only
to the main agent process during stop, under the assumption that
the agent will gracefully terminate its child processes. As
a backup measure against leaving the child processes running
when the agent terminates ungracefully, the watchdog sends a
SIGKILL to the entire process group upon exit.

The patch also increases the exit timeout to 10 seconds.

Change-Id: I943885fb8b650b8767ef0c6fa5c76bc91d987732
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>