FIP64: Allocate the NAT pool between multiple uplinks
This modifies the VPP state trait to track the group ports for
every uplink port, and use this information to allocate a partition
of a NAT pool whenever configuring a FIP64. The partition
allocation is deterministic, and therefore for the same uplink
topology, the trait returns the same allocated NAT pool for a
given FIP.

This implementation performs a crude mapping between the tenant
router port where a FIP64 is configured and the uplink ports,
under the following assumptions:

* For a given tenant router port where a FIP64 is associated,
that port is able to reach any of the configured uplink ports,
and therefore the NAT pool allocation should consider that port.

* We support only one uplink port per physical gateway, and the
NAT pool allocation algorithm considers the stateful port groups
of that port alone. The NAT pool will be deterministically divided
between all the ports in the stateful port groups of that uplink
port.

* If there is no uplink port, or if there are more than one, we
don't support the NAT pool allocation.

Ref: MI-1783

Change-Id: I5a010b413feea86b71494a082ccff756d85f484c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>