Fix bug where DSCP mark is set incorrectly
The DSCP mark field was being set into the entire TOS field,
however, the DSCP mark field is really the top 6 bits of the
TOS field, with the lower 2 bits being ECN.  So, we have to shift
the DSCP value by 2, then add in the original ECN value (just in
case it was set).

The test needed to be updated to expect this behavior.

Change-Id: Ia96841e15cb0e1427021839e9ddf85feee40079c