Conveyor belt receives an execution context
We are using the conveyor belt with a potential pitfall. It is
not thread-safe and it requires that the thread completing a
work item be the same as the thread submitting them.

However, it is possible in its usage that futures completing
outstanding work items to complete on different thread. When this
is the case the result of those futures should be rescheduled to
the thread that is submitting the work, but it is easy to overlook
this restriction.

To guarantee that the conveyor belt remains thread-safe, this
change adds to it an execution context argument, that allows the
task completion to be handled on a thread of the caller's choice
even if the task itself completes on a different thread. This
requires the users of the conveyor belt to provide an execution
context that is the same as the one submitting work items.

Change-Id: I20ebbdfcc1286b047e607ff9563084242c060544
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>