Update patch set 2
Patch Set 2:

The problem is that the queue inside the conveyor is not concurrent. This queue is accessed in two places:

1. When adding a new task to the conveyor belt
2. When completing a current task and fetching the next task to execute.

Since the queue is not concurrent, both these action need to happen on the same thread. While it's easy to control the thread on which we add tasks to the conveyor belt (e.g. in the VppController this is our single threaded executor), the completion of a task normally is done on the thread asynchronously performing that task (e.g. the thread running a command in background or talking to the VPP API).

Since originally the conveyor belt used a calling thread execution context to complete a task and fetch one from the queue, it meant that this would happen on the thread that completes the task future.

This is wrong, and until now we were careless about the restriction that futures returned by conveyor belt task should always complete on the thread that added the tasks to the queue in the first place.

Since this is an easy mistake to make, instead of modifying all our futures to make sure they complete on the correct thread, it's easier to just modify the conveyor belt, and instead use the calling thread to complete a task and access the next one from the queue, we just say on which thread this completion should happen.

Patch-set: 2