FIP64: Add proper cleanup to IPv6 interfaces
In VPP, an interface that has had an IPv6 address needs to be
cleaned up before removal. Otherwise the internal state of VPP
will be inconsistent, with dangling references to the removed
interface, and will eventually lead to a crash.

This cleanup is performed by the sw_interface_ip6_enable_disable
API call, equivalent to `disable ip6 interface NNN` in the
command-line. It removes the interface from neighbour discovery
tables.

Change-Id: Idf81465d68e1d0a7ae8246e9e9e199c02f5b4f7b
Signed-off-by: Adrian Serrano <adrian@midokura.com>