Add field to pool object to check if lbaasv2
The original implementation of L4LB heath monitoring deploys a
namespace with haproxy running inside to monitor members. It deploys
all of this directly from an Actor running in the agent. However,
for LBaaSv2 service containers are used to deploy this namespace.

The problem that this fix addresses is the fact that both the
service containers and the original implementation listen to the same
mapper (PoolHealthMonitorMapper) and so for a single deployement of an
LBaaSv2 topology two namespaces will be deployed.

This patch addresses this problem by adding a new field to the pool
object definition that will serve as a way to differentiate between a
pool created for the old implementation, and a pool created through
LBaaSv2.

In the LBaasV2 pool translator we set this new field to "true", and
the original implementation is modified to exclude all of these mappings
from the set it attempts to deploy.

This change is not intended as a permanent solution, and in the future
when we migrate the original implementation to use service containers,
the field will no longer be needed.

Change-Id: I33b0a9ef0cf1fda1822200e9165231d613e38500
Signed-off-by: Joe Mills <joe@midokura.com>