Use midonet data models for getFipRtrPortId
There was a bug in updating the subnet list of Neutron Network model
prior to 5.4, which was fixed[1] but the old incorrect data remained,
which prevents FIP deletion to work properly.  To get around this,
instead of NeutronNetwork and NeutronSubnet, use their MidoNet
counterparts, Network and Dhcp in getFipRtrPortId method.

This also takes a step in ensuring deletion FIPs created prior to v5.4
will succeed, which is likely required in v5.X -> v5.4 FIP data
migration.

[1] https://review.gerrithub.io/#/c/302557/
Ref: MI-1617
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>

Change-Id: Id7b32e96f456cad2f6a68f33e9dc08ebb22e6132
Signed-off-by: Mateo Hermida <mateo@midokura.com>