Socket native method throws last error exception
This changes the native socket method to throw a last error
exception on failure and force the calling code to handle the
exception rather than simply logging and continue. Not handling
the socket error will cause a null-pointer exception, and hide
the real cause of the failure.

The handling of the socket exception involves modifying the
Netlink and UNIX channels, and their corresponding providers
to throw an IOException whenever a failure occurs during the
channel initialization. Normally callers of these channel
providers either handle the IOException and wrap it and re-throw
it.

Ref: MI-1858

Change-Id: If6da8671d9ebe920d6504a9a549bf5b93cfe5c9c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit ba993fba460969baa0bbba6e08dd61a19e99732d)