VPP controller disables csum offloading
Currently the FIP64 feature does not support TCP/UDP checksum
offloading on the downlink tunnel interface. Having checksum
offloading enabled may cause packets from instances to reach
their destinations with an incorrect checksum when they originate
on the same host as the FIP64 gateway. The reason for this is
that there is no physical interface between the VM and the
VPP pipeline that will compute the checksum of the VXLAN-encap-
sulated TCP/UDP segments.

With this change TCP/UDP segments encapsulated in VXLAN and
bound for the VPP pipeline will have their checksum computed when
reaching the virtual downlink tunnel interface and before
being forwarded to VPP.

Ref: MI-1858

Change-Id: I13841bca712deeae555fb3274733b19629a5cc6a
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>