VppSetup logs additional parameters
This allows the VPP setup to log additional setup parameters for
troubleshooting purposes. To this end, the patch replaces the
task name property from a string to a task name function that is
only called when a particular logging level is enabled.

Additional assorted changes includes indentation, comments and
replacing selected objects allocations with singletons.

Change-Id: I30b729945f53be2a2522ca4867df3d1f69b55494
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>