Minor fixes to VppSetup
This includes indentation, logging, comments and replacing
selected objects allocations with singletons. With this change
logging messages include additional information on each setup
step. To ensure strings are not generated until the corresponding
logging level is enabled, and to include fields initialized after
the creation of the setup instances, the tasks name are replaced
with functions instead of strings.

Change-Id: I30b729945f53be2a2522ca4867df3d1f69b55494
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>