Remove failed postponed contexts from waiting room
This fixes an issue where a failure on a packet context that
has been postponed previously and added to the waiting room,
is kept in the waiting room even if the packet context is
return to the packet context pool when the failure occurs.

This will cause the simulation to try to handle an invalid
packet context later on, when the context expires in the waiting
room.

Change-Id: Ib83a742ea35fde23a80a127fd7056194fbb2f2e1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>