Introduce VPP FIP64 trait
This trait manages the FIP64 topology for the VPP controller. In
the initial implementation, this trait will maintain the state
for the provider routers corresponding to local uplink ports
and update their gateway table. This gateway table is subsequently
used by compute nodes to forward FIP64 traffic egressing from
a particular tenant router.

Ref: MI-1897

Change-Id: I6d2f4fa1f74c35274453676ba62d0d2ed886999c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>