Fix flake in TestMonitoredDaemonProcess
This test flakes because of unsynchronized access to a linked list
which is shared by multiple threads. It only needs to be accessed by
tests, so this patch makes the list private, and add an accessor that
copies the list.

Ref MI-1886

Change-Id: I5bf826b14c5b9006f3b2f8e3fb8c9cf422952014
Signed-off-by: Ivan Kelly <ivan@midokura.com>
(cherry picked from commit e70a56970af77866f29a27f5af6cc4d350094e4c)