Allow IPv6 over L2
This change allows the agent to simulate IPv6 packets ingressing
and traversing L2 ports (bridge and VXLAN ports). For this, it
allows the PacketWorkflow to process all packets, including IPv6,
and router and service ports will discard those packets on
ingress. The simulation will mark the packet as userspace to
prevent the installation of a drop flow in OVS, that can
interfere with the VPP controller.

Ref: MI-1692

Change-Id: I646d6f2636eeb7ca89fa9c82f1ff787b20d938e4
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>