Fip64 route type
Rather than encoding the fip64 nat range in a route and in a rule, just
encoding it in a route and add special handling at the router to mark
the packet for translation. This removes redundant processing that was
happening at both the router and the router port.

Ref MI-1927

Change-Id: I778e88a05e0fecf09ebe2592d2edd2d17cb000ca
Signed-off-by: Ivan Kelly <ivan@midokura.com>
(cherry picked from commit 9fad8fef12748eccfd5a142d8e450068c4ebf9cd)