FIP64: VPP state uses topology to send flow state
The gateways table will be scoped by external network. To this
end, the VPP controller no longer to needs to monitor this table
to determine to which hosts it must forward the flow state.
Instead it will rely on the topology data provided the by the
FIP64 topology trait, which monitors the uplink ports belonging
to the same stateful port group as the local uplink.

Ref: MI-1897

Change-Id: I5f143582820094c27661b89db1ac1d92448dfb81
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit edccc6f563f87d641559f3c0cf0753fd32a59543)