Use Scala foreach for Java iterables
This provides an implicit conversion from a Java iterable to an
AnyVal wrapper class that implements the foreach method. This
allows using scala for loops for Java iterables without allocating
a new object. This is unlike the scala.collections Java
conversion methods that allocate a new wrapper object for any
conversion.

Change-Id: I93e4a5da4e83e962adc32800150f9ba59f710087
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>