FIP64: Scope FIP64 table by external network
This modifies the FIP64 state table to be scoped by external
network instead of a global state table. To this end, the
VppDownlink trait is changed to monitor tables for every external
network reported by the FIP64 topology, and provide notifications
when FIP64 entries are added to any of those tables.

Change-Id: Ia05db7d26c2227965d949a2747c6de4b556e881c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>