Storage observables fail on ZK disconnect
This fixes an issue in the NodeObservable, PathDirectoryObservable
and ScalableStateTableManager classes, where the stream of
notifications would complete with an error if a ZK operation
would fail with a CONNECTIONLOSS error code. Instead, this should
have been a SESSIONEXPIRED error, and the implementation should
simply ignore CONNECTIONLOSS. Recovery upon reconnection is
already supported.

Change-Id: I8b43249a1ad5fe149f064113bf7a68bccf192d9c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>