Remove the interfaces explicitly on bgpd_helper down
It seems that under some circumstances, veth interfaces are not
properly clean for BGP. This could happen if the bgpd_helper is
interrupted before adding them to the namespace, since the current
code removes the namespace and rely on cascade removal of
interfaces, that usually are added correctly to the namespace.

This change guard against this odd situation, by explicitly
removing the interfaces, regardless of them belonging to the
namespace.

Ref: MI-1904

Change-Id: I49f617c8e494edcee6bec0cdf68255e1ec3261d5
Signed-off-by: Miguel Herranz <miguel@midokura.com>