FIP64 table removes unused field and adds subnets
This removes the unused router identifer from the FIP64 table
entries in order to make the entries shorter and require less
data. Also, the patch adds the subnet network length to both the
floating and the fixed IP to be available to the VPP controller
if needed. The fixed IP uses a default length 32, this can be
updated later if used.

Ref: MI-1860

Change-Id: I5def460760558d3fcfe9910759d312d53af4fde5
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>