FIP64: Improve VPP threading and sequentiality
This strengthens the VPP controller threading and sequentiality
guarantees, by ensuring that calls that involve virtual
topology observables are executed on the virtual topology
thread, and all subsequent notifications are also emitted on
the same thread. This fixes an issue in the VppProviderRouter
class where a call from the VPP controller would trigger an RX
notification that was not scheduled on the VT thread. This will
subsequently cause a thread check assertion to fail, terminating
the observable stream.

In addition, this modifies the logic from the FIP64 topology
such that certain steps that require the completion of setup
tasks by the VPP controller, await for the successful completion
of those tasks before proceeding. This ensures that the VPP
controller will not receive notifications inconsistent with its
own setup logic.

Change-Id: I9c99c06c79fbf2e1b04ab465e258ec1cc11c206a
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>