LBV2: Agent reads PoolMember status from StateStore
Update agent to read PoolMember status from StateStorage when the status in the
PoolMember protobuf object is set to MONITORED, as it always will be for LBaaS
V2 PoolMembers. When status cannot be found in state storage, the PoolMember is
treated as unmonitored and assumed to be active.

Change-Id: I13fbe0c8abf311e8094b950f680214db31eb32d0
Signed-off-by: Brandon Berg <bberg@midokura.com>
(cherry picked from commit 908bf1fc30175b9b234d98f019004e9e727ca7fe)