Backend caches state tables
The state proxy client does not support multiple subscribers
for the same state table because this would imply that the client
must cache the table data itself, in order to provide a consistent
stream of updates for all subscribers. Instead, this funcionality
exists in the ScalableStateTable, which is acts as a subscriber
to the state proxy client and allows concurrent access for
multiple users, including read/write operations, subscription
to notifications, start/stopping the table, etc.

For this reason, there can only be a single ScalableStateTable
instance for a given state table, and the backend must ensure
that multiple requests for the same do not create more that one
instance.

To this end, this patch changes the ZookeeperStateTable trait
to provide caching for state tables.

Fixes: MI-1946

Change-Id: I368fe92bd99a3da2c353ac5824ecedded5d62bbd
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>