LBV2: Don't cascade Pool deletes to VIPs
Deletion of a Pool via the LBV2 API was incorrectly cascading to the
VIPs associated with the Pool. The binding is also used by Midonet API,
and changing it might cause breaking changes, so instead I'm just
adding an update operation to clear the pool's VIP IDs before deleting
it in LBV2. Since the vipIds list is empty, the delete won't cascade.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-1950
Change-Id: Ic98c5d91d966d7dcc5445f2bc7dca4226233f72f